+++ src/friendly_computing_machine/bot/event_handler.py
@@ -2,10 +2,6 @@ import datetime
 import logging
 
 from friendly_computing_machine.bot.app import app, get_bot_config
-from friendly_computing_machine.bot.workflow import (
-    SlackConextGeminiWorkflow,
-    SlackContextGeminiWorkflowParams,
-)
 from friendly_computing_machine.db.dal import (
     insert_genai_text,
     insert_slack_command,
@@ -16,6 +12,10 @@ from friendly_computing_machine.models.genai import GenAITextCreate
 from friendly_computing_machine.models.slack import (
     SlackCommandCreate,
     SlackMessageCreate,
+)
+from friendly_computing_machine.temporal.slack.workflow import (
+    SlackConextGeminiWorkflow,
+    SlackContextGeminiWorkflowParams,
 )
 from friendly_computing_machine.temporal.util import (
     execute_workflow,
