+++ src/friendly_computing_machine/temporal/ai/activity.py
