+++ src/friendly_computing_machine/temporal/db/job_activity.py
