+++ src/friendly_computing_machine/temporal/slack/activity.py
