+++ src/friendly_computing_machine/temporal/worker.py
@@ -8,17 +8,12 @@ from temporalio.worker.workflow_sandbox import (
     SandboxRestrictions,
 )
 
-from friendly_computing_machine.bot.activity import (
-    backfill_slack_user_info_activity,
-    generate_context_prompt,
-    get_slack_channel_context,
+from friendly_computing_machine.temporal.ai.activity import (
+    generate_gemini_response,
+    generate_summary,
 )
-from friendly_computing_machine.bot.workflow import (
-    SlackConextGeminiWorkflow,
-    SlackMessageQODWorkflow,
-    SlackUserInfoWorkflow,
-)
-from friendly_computing_machine.db.job_activity import (
+from friendly_computing_machine.temporal.base import AbstractScheduleWorkflow
+from friendly_computing_machine.temporal.db.job_activity import (
     backfill_genai_text_slack_channel_id_activity,
     backfill_genai_text_slack_user_id_activity,
     backfill_slack_messages_slack_channel_id_activity,
@@ -28,15 +23,20 @@ from friendly_computing_machine.db.job_activity import (
     delete_slack_message_duplicates_activity,
     upsert_slack_user_creates_activity,
 )
-from friendly_computing_machine.gemini.activity import (
-    generate_gemini_response,
-    generate_summary,
-)
-from friendly_computing_machine.temporal.base import AbstractScheduleWorkflow
 from friendly_computing_machine.temporal.sample import (
     SayHello,
     build_hello_prompt,
     say_hello,
+)
+from friendly_computing_machine.temporal.slack.activity import (
+    backfill_slack_user_info_activity,
+    generate_context_prompt,
+    get_slack_channel_context,
+)
+from friendly_computing_machine.temporal.slack.workflow import (
+    SlackConextGeminiWorkflow,
+    SlackMessageQODWorkflow,
+    SlackUserInfoWorkflow,
 )
 from friendly_computing_machine.temporal.util import (
     get_temporal_client_async,
